RReorganización, usuarios y roles
Se redefinieron las tablas de la base de datos. Además se añadieron las tablas necesarias para roles y usuarios. Estos cambios se ven reflejados en los modelos, que quedaron acorde a las respectivas migraciones

+++ database/migrations/2019_01_04_000000_create_ventas_table.php
@@ -14,11 +14,11 @@ class CreateVentasTable extends Migration
     public function up()
     {
         Schema::create('ventas', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->date('fecha');
             $table->string('direccion', 40)->nullable();
-            $table->unsignedBigInteger('cliente_id');
-            $table->unsignedBigInteger('vendedor_id');
+            $table->unsignedInteger('cliente_id');
+            $table->unsignedInteger('vendedor_id');
             $table->double('valor_total', 8, 2)->default(0.0);
             $table->unsignedInteger('estado_venta_id');
             $table->softDeletes();
